requirements.txt
pandas
scikit-learn
seaborn
matplotlib
streamlit
python-dotenv
